+++ examples/with-clerk-appdir/src/app/layout.tsx
@@ -3,6 +3,11 @@ import "@uploadthing/react/styles.css";
 import { Inter } from "next/font/google";
 import { ClerkProvider } from "@clerk/nextjs";
 
+import { NextSSRPlugin } from "@uploadthing/react/next-ssr-plugin";
+import { extractRouterConfig } from "uploadthing/server";
+
+import { uploadRouter } from "~/server/uploadthing";
+
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata = {
@@ -17,7 +22,10 @@ export default function RootLayout({
   return (
     <ClerkProvider>
       <html lang="en">
-        <body className={inter.className}>{children}</body>
+        <body className={inter.className}>
+          <NextSSRPlugin routerConfig={extractRouterConfig(uploadRouter)} />
+          {children}
+        </body>
       </html>
     </ClerkProvider>
   );