+++ examples/with-drizzle-pagesdir/src/pages/_app.tsx
@@ -1,0 +1,9 @@
+import type { AppProps } from "next/app";
+
+import "@uploadthing/react/styles.css";
+
+function MyApp({ Component, pageProps }: AppProps) {
+  return <Component {...pageProps} />;
+}
+
+export default MyApp;