+++ packages/react/src/components/button.tsx
@@ -1,25 +1,23 @@
-import { useRef, useState } from "react";
+import { useEffect, useRef, useState } from "react";
 import { twMerge } from "tailwind-merge";
 
-import { generateMimeTypes } from "uploadthing/client";
+import {
+  allowedContentTextLabelGenerator,
+  contentFieldToContent,
+  generateMimeTypes,
+  generatePermittedFileTypes,
+  styleFieldToClassName,
+  styleFieldToCssObject,
+} from "uploadthing/client";
+import type { ContentField, StyleField } from "uploadthing/client";
 import type { ErrorMessage, FileRouter } from "uploadthing/server";
 
 import type { UploadthingComponentProps } from "../types";
 import { INTERNAL_uploadthingHookGen } from "../useUploadThing";
-import type { ContentField, StyleField } from "../utils/styles";
-import {
-  contentFieldToContent,
-  styleFieldToClassName,
-  styleFieldToCssObject,
-} from "../utils/styles";
-import {
-  allowedContentTextLabelGenerator,
-  generatePermittedFileTypes,
-  progressWidths,
-  Spinner,
-} from "./shared";
+import { getFilesFromClipboardEvent, progressWidths, Spinner } from "./shared";
 
 type ButtonStyleFieldCallbackArgs = {
+  __runtime: "react";
   ready: boolean;
   isUploading: boolean;
   uploadProgress: number;
@@ -32,12 +30,18 @@ export type UploadButtonProps<TRouter extends FileRouter> =
       container?: StyleField<ButtonStyleFieldCallbackArgs>;
       button?: StyleField<ButtonStyleFieldCallbackArgs>;
       allowedContent?: StyleField<ButtonStyleFieldCallbackArgs>;
+      clearBtn?: StyleField<ButtonStyleFieldCallbackArgs>;
     };
     content?: {
       button?: ContentField<ButtonStyleFieldCallbackArgs>;
       allowedContent?: ContentField<ButtonStyleFieldCallbackArgs>;
+      clearBtn?: ContentField<ButtonStyleFieldCallbackArgs>;
     };
     className?: string;
+    config?: {
+      appendOnPaste?: boolean;
+      mode?: "auto" | "manual";
+    };
   };
 
 /**
@@ -66,14 +70,22 @@ export function UploadButton<TRouter extends FileRouter>(
     // Allow to disable the button
     __internal_button_disabled?: boolean;
   };
+
+  const { mode = "auto", appendOnPaste = false } = $props.config ?? {};
+
   const useUploadThing = INTERNAL_uploadthingHookGen<TRouter>();
 
   const fileInputRef = useRef<HTMLInputElement>(null);
+  const labelRef = useRef<HTMLLabelElement>(null);
   const [uploadProgressState, setUploadProgress] = useState(
     $props.__internal_upload_progress ?? 0,
   );
+  const [files, setFiles] = useState<File[]>([]);
+  const [isManualTriggerDisplayed, setIsManualTriggerDisplayed] =
+    useState(false);
   const uploadProgress =
     $props.__internal_upload_progress ?? uploadProgressState;
+
   const { startUpload, isUploading, permittedFileInfo } = useUploadThing(
     $props.endpoint,
     {
@@ -81,6 +93,8 @@ export function UploadButton<TRouter extends FileRouter>(
         if (fileInputRef.current) {
           fileInputRef.current.value = "";
         }
+        setIsManualTriggerDisplayed(false);
+        setFiles([]);
         $props.onClientUploadComplete?.(res);
         setUploadProgress(0);
       },
@@ -101,32 +115,63 @@ export function UploadButton<TRouter extends FileRouter>(
     $props.__internal_ready ??
     ($props.__internal_state === "ready" || fileTypes.length > 0);
 
+  useEffect(() => {
+    const handlePaste = (event: ClipboardEvent) => {
+      if (!appendOnPaste) return;
+      if (document.activeElement !== labelRef.current) return;
+
+      const pastedFiles = getFilesFromClipboardEvent(event);
+      if (!pastedFiles) return;
+
+      setFiles((prev) => [...prev, ...pastedFiles]);
+
+      if (mode === "auto") {
+        const input = "input" in $props ? $props.input : undefined;
+        void startUpload(files, input);
+      }
+    };
+
+    window.addEventListener("paste", handlePaste);
+    return () => {
+      window.removeEventListener("paste", handlePaste);
+    };
+  }, [startUpload, appendOnPaste, $props, files, mode, fileTypes]);
+
   const getUploadButtonText = (fileTypes: string[]) => {
+    if (isManualTriggerDisplayed)
+      return `Upload ${files.length} file${files.length === 1 ? "" : "s"}`;
     if (fileTypes.length === 0) return "Loading...";
     return `Choose File${multiple ? `(s)` : ``}`;
   };
 
   const getInputProps = () => ({
-    className: "hidden",
     type: "file",
     ref: fileInputRef,
     multiple,
     accept: generateMimeTypes(fileTypes ?? [])?.join(", "),
     onChange: (e: React.ChangeEvent<HTMLInputElement>) => {
       if (!e.target.files) return;
+      const selectedFiles = Array.from(e.target.files);
+
+      if (mode === "manual") {
+        setFiles(selectedFiles);
+        setIsManualTriggerDisplayed(true);
+        return;
+      }
+
       const input = "input" in $props ? $props.input : undefined;
-      const files = Array.from(e.target.files);
-      void startUpload(files, input);
+      void startUpload(selectedFiles, input);
     },
     disabled: $props.__internal_button_disabled ?? !ready,
+    ...(!($props.__internal_button_disabled ?? !ready) ? { tabIndex: 0 } : {}),
   });
 
-  const styleFieldArg: ButtonStyleFieldCallbackArgs = {
+  const styleFieldArg = {
     ready: ready,
     isUploading: $props.__internal_state === "uploading" || isUploading,
     uploadProgress,
     fileTypes,
-  };
+  } as ButtonStyleFieldCallbackArgs;
 
   const state = (() => {
     if ($props.__internal_state) return $props.__internal_state;
@@ -135,6 +180,47 @@ export function UploadButton<TRouter extends FileRouter>(
 
     return "uploading";
   })();
+
+  const renderClearButton = () => (
+    <button
+      onClick={() => {
+        setFiles([]);
+        setIsManualTriggerDisplayed(false);
+
+        if (fileInputRef.current) {
+          fileInputRef.current.value = "";
+        }
+      }}
+      className={twMerge(
+        "h-[1.25rem] cursor-pointer rounded border-none bg-transparent text-gray-500 transition-colors hover:bg-slate-200 hover:text-gray-600",
+        styleFieldToClassName($props.appearance?.clearBtn, styleFieldArg),
+      )}
+      style={styleFieldToCssObject($props.appearance?.clearBtn, styleFieldArg)}
+      data-state={state}
+      data-ut-element="clear-btn"
+    >
+      {contentFieldToContent($props.content?.clearBtn, styleFieldArg) ??
+        "Clear"}
+    </button>
+  );
+
+  const renderAllowedContent = () => (
+    <div
+      className={twMerge(
+        "h-[1.25rem]  text-xs leading-5 text-gray-600",
+        styleFieldToClassName($props.appearance?.allowedContent, styleFieldArg),
+      )}
+      style={styleFieldToCssObject(
+        $props.appearance?.allowedContent,
+        styleFieldArg,
+      )}
+      data-state={state}
+      data-ut-element="allowed-content"
+    >
+      {contentFieldToContent($props.content?.allowedContent, styleFieldArg) ??
+        allowedContentTextLabelGenerator(permittedFileInfo?.config)}
+    </div>
+  );
 
   return (
     <div
@@ -148,18 +234,28 @@ export function UploadButton<TRouter extends FileRouter>(
     >
       <label
         className={twMerge(
-          "relative flex h-10 w-36 cursor-pointer items-center justify-center overflow-hidden rounded-md text-white after:transition-[width] after:duration-500",
+          "relative flex h-10 w-36 cursor-pointer items-center justify-center overflow-hidden rounded-md text-white after:transition-[width] after:duration-500 focus-within:ring-2 focus-within:ring-blue-600 focus-within:ring-offset-2",
           state === "readying" && "cursor-not-allowed bg-blue-400",
           state === "uploading" &&
-            `bg-blue-400 after:absolute after:left-0 after:h-full after:bg-blue-600 ${progressWidths[uploadProgress]}`,
+            `bg-blue-400 after:absolute after:left-0 after:h-full after:bg-blue-600 after:content-[''] ${progressWidths[uploadProgress]}`,
           state === "ready" && "bg-blue-600",
           styleFieldToClassName($props.appearance?.button, styleFieldArg),
         )}
         style={styleFieldToCssObject($props.appearance?.button, styleFieldArg)}
         data-state={state}
         data-ut-element="button"
+        tabIndex={0}
+        ref={labelRef}
+        onClick={(e) => {
+          if (isManualTriggerDisplayed) {
+            e.preventDefault();
+            e.stopPropagation();
+            const input = "input" in $props ? $props.input : undefined;
+            void startUpload(files, input);
+          }
+        }}
       >
-        <input {...getInputProps()} />
+        <input {...getInputProps()} className="sr-only" />
         {contentFieldToContent($props.content?.button, styleFieldArg) ??
           (state === "uploading" ? (
             <Spinner />
@@ -167,24 +263,9 @@ export function UploadButton<TRouter extends FileRouter>(
             getUploadButtonText(fileTypes)
           ))}
       </label>
-      <div
-        className={twMerge(
-          "h-[1.25rem]  text-xs leading-5 text-gray-600",
-          styleFieldToClassName(
-            $props.appearance?.allowedContent,
-            styleFieldArg,
-          ),
-        )}
-        style={styleFieldToCssObject(
-          $props.appearance?.allowedContent,
-          styleFieldArg,
-        )}
-        data-state={state}
-        data-ut-element="allowed-content"
-      >
-        {contentFieldToContent($props.content?.allowedContent, styleFieldArg) ??
-          allowedContentTextLabelGenerator(permittedFileInfo?.config)}
-      </div>
+      {mode === "manual" && files.length > 0
+        ? renderClearButton()
+        : renderAllowedContent()}
     </div>
   );
 }