+++ packages/react/src/types.ts
@@ -1,4 +1,7 @@
-import type { UploadThingError } from "@uploadthing/shared";
+import type {
+  ExpandedRouteConfig,
+  UploadThingError,
+} from "@uploadthing/shared";
 import type { UploadFileResponse } from "uploadthing/client";
 import type {
   FileRouter,
@@ -24,3 +27,8 @@ export type UploadthingComponentProps<TRouter extends FileRouter> = {
         input: inferEndpointInput<TRouter[TEndpoint]>;
       });
 }[keyof TRouter];
+
+export type EndpointMetadata = {
+  slug: string;
+  config: ExpandedRouteConfig;
+}[];