+++ packages/react/src/useUploadThing.ts
@@ -1,6 +1,5 @@
 import { useRef, useState } from "react";
 
-import type { ExpandedRouteConfig } from "@uploadthing/shared";
 import { UploadThingError } from "@uploadthing/shared";
 import type { UploadFileResponse } from "uploadthing/client";
 import { DANGEROUS__uploadFiles } from "uploadthing/client";
@@ -11,17 +10,21 @@ import type {
   inferErrorShape,
 } from "uploadthing/server";
 
+import type { EndpointMetadata } from "./types";
 import { useEvent } from "./utils/useEvent";
 import useFetch from "./utils/useFetch";
 
-type EndpointMetadata = {
-  slug: string;
-  config: ExpandedRouteConfig;
-}[];
+declare const globalThis: {
+  __UPLOADTHING?: EndpointMetadata;
+};
 
 const useEndpointMetadata = (endpoint: string) => {
-  const { data } = useFetch<EndpointMetadata>("/api/uploadthing");
-  return data?.find((x) => x.slug === endpoint);
+  const maybeServerData = globalThis.__UPLOADTHING;
+  const { data } = useFetch<EndpointMetadata>(
+    // Don't fetch if we already have the data
+    maybeServerData ? undefined : "/api/uploadthing",
+  );
+  return (maybeServerData ?? data)?.find((x) => x.slug === endpoint);
 };
 
 export type UseUploadthingProps<
@@ -62,6 +65,7 @@ export const INTERNAL_uploadthingHookGen = <TRouter extends FileRouter>() => {
     const startUpload = useEvent(async (...args: FuncInput) => {
       const [files, input] = args;
       setUploading(true);
+      opts?.onUploadProgress?.(0);
       try {
         const res = await DANGEROUS__uploadFiles<TRouter, TEndpoint>(endpoint, {
           files,